requirements.txt
asgiref==3.8.1
cairocffi==1.7.1
CairoSVG==2.8.2
cffi==1.17.1
cssselect2==0.8.0
defusedxml==0.7.1
Django==5.2.3
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-authtoken==3.16.0
multiavatar==1.0.3
pillow==11.2.1
pycparser==2.22
sqlparse==0.5.3
tinycss2==1.4.0
tzdata==2025.2
webencodings==0.5.1
